refactor: Don't rearrange array keys

+++ app/classes/Cookies.php
@@ -88,15 +88,8 @@ class Cookies {
         // Remove cookie item from the array
         unset($this->cookie_data[$id]);
 
-        // Sort array with a new key values
-        $this->cookie_data = array_values($this->cookie_data);
-
-        if (empty($this->cookie_data)) {
-            $this->delete();
-            return;
-        }
-
-        // Update cookie
         $this->update();
+
+        if (empty($this->cookie_data)) $this->delete();
     }
 }
